refactor: 오목 패키지 분리
