--- src/main/kotlin/ColumnPosition.kt
@@ -1,4 +1,0 @@
-@JvmInline
-value class ColumnPosition(
-    val value: Int,
-)
