--- src/main/kotlin/DrawRule.kt
@@ -1,11 +1,0 @@
-class DrawRule : OmokRule {
-    override fun canPlace(
-        omokBoard: OmokBoard,
-        playerStone: PlayerStone,
-    ): PlaceResult =
-        if (omokBoard.value.all { it.state != PointState.EMPTY }) {
-            PlaceResult.Success.Finish(GameResult.DRAW)
-        } else {
-            PlaceResult.Success.Progress
-        }
-}
