--- src/main/kotlin/GameResult.kt
@@ -1,5 +1,0 @@
-enum class GameResult {
-    WIN_BLACK,
-    WIN_WHITE,
-    DRAW,
-}
