--- src/main/kotlin/OmokRule.kt
@@ -1,6 +1,0 @@
-interface OmokRule {
-    fun canPlace(
-        omokBoard: OmokBoard,
-        playerStone: PlayerStone,
-    ): PlaceResult
-}
