--- src/main/kotlin/PlayerStone.kt
@@ -1,4 +1,0 @@
-data class PlayerStone(
-    val color: StoneColor,
-    val position: Position,
-)
