--- src/main/kotlin/PointState.kt
@@ -1,5 +1,0 @@
-enum class PointState {
-    OCCUPIED_BLACK,
-    OCCUPIED_WHITE,
-    EMPTY,
-}
