--- src/main/kotlin/Position.kt
@@ -1,4 +1,0 @@
-data class Position(
-    val row: RowPosition,
-    val column: ColumnPosition,
-)
