--- src/main/kotlin/RowPosition.kt
@@ -1,4 +1,0 @@
-@JvmInline
-value class RowPosition(
-    val value: Int,
-)
