--- src/main/kotlin/StoneColor.kt
@@ -1,11 +1,0 @@
-enum class StoneColor {
-    BLACK,
-    WHITE,
-    ;
-
-    fun reverse(): StoneColor =
-        when (this) {
-            BLACK -> WHITE
-            WHITE -> BLACK
-        }
-}
