+++ src/main/kotlin/omok/model/domain/omokboard/ColumnPosition.kt
@@ -1,0 +1,6 @@
+package omok.model.domain.omokboard
+
+@JvmInline
+value class ColumnPosition(
+    val value: Int,
+)
