+++ src/main/kotlin/omok/model/domain/omokboard/RowPosition.kt
@@ -1,0 +1,6 @@
+package omok.model.domain.omokboard
+
+@JvmInline
+value class RowPosition(
+    val value: Int,
+)
