+++ src/main/kotlin/omok/model/domain/player/StoneColor.kt
@@ -1,0 +1,13 @@
+package omok.model.domain.player
+
+enum class StoneColor {
+    BLACK,
+    WHITE,
+    ;
+
+    fun reverse(): StoneColor =
+        when (this) {
+            BLACK -> WHITE
+            WHITE -> BLACK
+        }
+}
