+++ src/main/kotlin/omok/model/domain/rule/AlreadyExistRule.kt
@@ -1,0 +1,17 @@
+package omok.model.domain.rule
+
+import omok.model.domain.omokboard.OmokBoard
+import omok.model.domain.omokboard.PointState
+import omok.model.domain.player.PlayerStone
+
+class AlreadyExistRule : OmokRule {
+    override fun canPlace(
+        omokBoard: OmokBoard,
+        playerStone: PlayerStone,
+    ): PlaceResult =
+        if (omokBoard.value.find { it.position == playerStone.position }?.state == PointState.EMPTY) {
+            PlaceResult.Success.Progress
+        } else {
+            PlaceResult.Failure.AlreadyExist
+        }
+}
