+++ src/main/kotlin/omok/model/domain/rule/DrawRule.kt
@@ -1,0 +1,17 @@
+package omok.model.domain.rule
+
+import omok.model.domain.omokboard.OmokBoard
+import omok.model.domain.omokboard.PointState
+import omok.model.domain.player.PlayerStone
+
+class DrawRule : OmokRule {
+    override fun canPlace(
+        omokBoard: OmokBoard,
+        playerStone: PlayerStone,
+    ): PlaceResult =
+        if (omokBoard.value.all { it.state != PointState.EMPTY }) {
+            PlaceResult.Success.Finish(GameResult.DRAW)
+        } else {
+            PlaceResult.Success.Progress
+        }
+}
