+++ src/main/kotlin/omok/model/domain/rule/GameResult.kt
@@ -1,0 +1,7 @@
+package omok.model.domain.rule
+
+enum class GameResult {
+    WIN_BLACK,
+    WIN_WHITE,
+    DRAW,
+}
