+++ src/main/kotlin/omok/model/domain/rule/PlaceResult.kt
@@ -1,0 +1,17 @@
+package omok.model.domain.rule
+
+sealed class PlaceResult {
+    sealed class Success : PlaceResult() {
+        data object Progress : Success()
+
+        data class Finish(
+            val result: GameResult,
+        ) : Success()
+    }
+
+    sealed class Failure : PlaceResult() {
+        data object AlreadyExist : Failure()
+
+        data object InvalidPosition : Failure()
+    }
+}
