--- src/test/kotlin/Fixture.kt
@@ -1,18 +1,0 @@
-val POSITION_ONE_ONE = Position(RowPosition(1), ColumnPosition(1))
-
-val POSITION_ONE_TWO = Position(RowPosition(1), ColumnPosition(2))
-val POSITION_ONE_THREE = Position(RowPosition(1), ColumnPosition(3))
-val POSITION_ONE_FOUR = Position(RowPosition(1), ColumnPosition(4))
-val POSITION_ONE_FIVE = Position(RowPosition(1), ColumnPosition(5))
-
-val POSITION_TWO_ONE = Position(RowPosition(2), ColumnPosition(1))
-val POSITION_THREE_ONE = Position(RowPosition(3), ColumnPosition(1))
-val POSITION_FOUR_ONE = Position(RowPosition(4), ColumnPosition(1))
-val POSITION_FIVE_ONE = Position(RowPosition(5), ColumnPosition(1))
-
-val POSITION_TWO_TWO = Position(RowPosition(2), ColumnPosition(2))
-val POSITION_THREE_THREE = Position(RowPosition(3), ColumnPosition(3))
-val POSITION_FOUR_FOUR = Position(RowPosition(4), ColumnPosition(4))
-val POSITION_FIVE_FIVE = Position(RowPosition(5), ColumnPosition(5))
-
-val POSITION_ZERO_ZERO = Position(RowPosition(0), ColumnPosition(0))
