--- src/test/kotlin/PlayingBoardTest.kt
@@ -1,8 +1,0 @@
-import org.junit.jupiter.api.Test
-
-class PlayingBoardTest {
-    @Test
-    fun `오목판에서 원하는 위치에 돌을 놓는다`() {
-        // TODO
-    }
-}
