+++ src/test/kotlin/omok/model/domain/omokboard/PlayingBoardTest.kt
@@ -1,0 +1,10 @@
+package omok.model.domain.omokboard
+
+import org.junit.jupiter.api.Test
+
+class PlayingBoardTest {
+    @Test
+    fun `오목판에서 원하는 위치에 돌을 놓는다`() {
+        // TODO
+    }
+}
