+++ src/test/kotlin/omok/model/domain/omokboard/PointTest.kt
@@ -1,0 +1,44 @@
+package omok.model.domain.omokboard
+
+import omok.POSITION_ONE_ONE
+import omok.model.domain.omokboard.PointState.EMPTY
+import omok.model.domain.omokboard.PointState.OCCUPIED_BLACK
+import omok.model.domain.omokboard.PointState.OCCUPIED_WHITE
+import omok.model.domain.player.StoneColor
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.Test
+
+class PointTest {
+    @Test
+    fun `초기 state는 EMPTY다`() {
+        // given & when
+        val actual = Point(POSITION_ONE_ONE).state
+        val expected = EMPTY
+        // then
+        assertThat(actual).isEqualTo(expected)
+    }
+
+    @Test
+    fun `검은돌을 두면 state가 검정색으로 바뀐다`() {
+        // given
+        val point = Point(POSITION_ONE_ONE)
+        // when
+        point.updateState(StoneColor.BLACK)
+        val actual = point.state
+        val expected = OCCUPIED_BLACK
+        // then
+        assertThat(actual).isEqualTo(expected)
+    }
+
+    @Test
+    fun `흰돌을 두면 state가 흰색으로 바뀐다`() {
+        // given
+        val point = Point(POSITION_ONE_ONE)
+        // when
+        point.updateState(StoneColor.WHITE)
+        val actual = point.state
+        val expected = OCCUPIED_WHITE
+        // then
+        assertThat(actual).isEqualTo(expected)
+    }
+}
