+++ src/test/kotlin/omok/model/domain/rule/DrawRuleTest.kt
@@ -1,0 +1,41 @@
+package omok.model.domain.rule
+
+import omok.model.domain.omokboard.ColumnPosition
+import omok.model.domain.omokboard.PlayingBoard
+import omok.model.domain.omokboard.Position
+import omok.model.domain.omokboard.RowPosition
+import omok.model.domain.player.PlayerStone
+import omok.model.domain.player.StoneColor
+import omok.model.domain.rule.DrawRule
+import omok.model.domain.rule.GameResult
+import omok.model.domain.rule.PlaceResult
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+
+class DrawRuleTest {
+    private lateinit var playingBoard: PlayingBoard
+
+    @BeforeEach
+    fun setup() {
+        playingBoard = PlayingBoard()
+    }
+
+    @Test
+    fun `모든 칸이 채워지면 무승부를 반환한다`() {
+        // given
+        for (row in 1..15) {
+            for (col in 1..15) {
+                val stoneColor = if ((row + col) % 2 == 0) StoneColor.BLACK else StoneColor.WHITE
+                val playerStone = PlayerStone(stoneColor, Position(RowPosition(row), ColumnPosition(col)))
+                playingBoard.placeStone(playerStone)
+            }
+        }
+
+        val playerStone = PlayerStone(StoneColor.BLACK, Position(RowPosition(15), ColumnPosition(15)))
+        val actual = DrawRule().canPlace(playingBoard.board, playerStone)
+        val expected = PlaceResult.Success.Finish(GameResult.DRAW)
+
+        assertThat(actual).isEqualTo(expected)
+    }
+}
