+++ src/test/kotlin/omok/model/domain/rule/WinningRuleTest.kt
@@ -1,0 +1,117 @@
+package omok.model.domain.rule
+
+import omok.POSITION_FIVE_FIVE
+import omok.POSITION_FIVE_ONE
+import omok.POSITION_FOUR_FOUR
+import omok.POSITION_FOUR_ONE
+import omok.POSITION_ONE_FIVE
+import omok.POSITION_ONE_FOUR
+import omok.POSITION_ONE_ONE
+import omok.POSITION_ONE_THREE
+import omok.POSITION_ONE_TWO
+import omok.POSITION_THREE_ONE
+import omok.POSITION_THREE_THREE
+import omok.POSITION_TWO_ONE
+import omok.POSITION_TWO_TWO
+import omok.model.domain.omokboard.PlayingBoard
+import omok.model.domain.player.PlayerStone
+import omok.model.domain.player.StoneColor
+import omok.model.domain.rule.GameResult
+import omok.model.domain.rule.PlaceResult
+import omok.model.domain.rule.WinningRule
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+
+class WinningRuleTest {
+    private lateinit var playingBoard: PlayingBoard
+
+    @BeforeEach
+    fun setup() {
+        playingBoard = PlayingBoard()
+    }
+
+    @Test
+    fun `같은 색의 돌이 가로로 5개 있으면 Finish를 반환한다`() {
+        // given
+        val playerStone1: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_ONE_ONE)
+        val playerStone2: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_ONE_TWO)
+        val playerStone3: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_ONE_THREE)
+        val playerStone4: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_ONE_FOUR)
+        val playerStone5: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_ONE_FIVE)
+
+        playingBoard.placeStone(playerStone1)
+        playingBoard.placeStone(playerStone2)
+        playingBoard.placeStone(playerStone3)
+        playingBoard.placeStone(playerStone4)
+        playingBoard.placeStone(playerStone5)
+
+        val actual = WinningRule().canPlace(playingBoard.board, playerStone5)
+        val expected = PlaceResult.Success.Finish(GameResult.WIN_BLACK)
+
+        assertThat(actual).isEqualTo(expected)
+    }
+
+    @Test
+    fun `같은 색의 돌이 세로로 5개 있으면 Finish한다`() {
+        // given
+        val playerStone1: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_ONE_ONE)
+        val playerStone2: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_TWO_ONE)
+        val playerStone3: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_THREE_ONE)
+        val playerStone4: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_FOUR_ONE)
+        val playerStone5: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_FIVE_ONE)
+
+        playingBoard.placeStone(playerStone1)
+        playingBoard.placeStone(playerStone2)
+        playingBoard.placeStone(playerStone3)
+        playingBoard.placeStone(playerStone4)
+        playingBoard.placeStone(playerStone5)
+
+        val actual = WinningRule().canPlace(playingBoard.board, playerStone5)
+        val expected = PlaceResult.Success.Finish(GameResult.WIN_BLACK)
+
+        assertThat(actual).isEqualTo(expected)
+    }
+
+    @Test
+    fun `같은 색의 돌이 대각선으로 5개 있으면 Finish를 반환한다`() {
+        // given
+        val playerStone1: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_ONE_ONE)
+        val playerStone2: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_TWO_TWO)
+        val playerStone3: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_THREE_THREE)
+        val playerStone4: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_FOUR_FOUR)
+        val playerStone5: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_FIVE_FIVE)
+
+        playingBoard.placeStone(playerStone1)
+        playingBoard.placeStone(playerStone2)
+        playingBoard.placeStone(playerStone3)
+        playingBoard.placeStone(playerStone4)
+        playingBoard.placeStone(playerStone5)
+
+        val actual = WinningRule().canPlace(playingBoard.board, playerStone5)
+        val expected = PlaceResult.Success.Finish(GameResult.WIN_BLACK)
+
+        assertThat(actual).isEqualTo(expected)
+    }
+
+    @Test
+    fun `나란히 있는 같은 색의 돌 4개 사이에 한 칸을 띄우고, 빈 공간에 돌을 두면 Finish를 반환한다`() {
+        // given
+        val playerStone1: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_ONE_ONE)
+        val playerStone2: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_ONE_TWO)
+        val playerStone3: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_ONE_THREE)
+        val playerStone4: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_ONE_FOUR)
+        val playerStone5: PlayerStone = PlayerStone(StoneColor.BLACK, POSITION_ONE_FIVE)
+
+        playingBoard.placeStone(playerStone1)
+        playingBoard.placeStone(playerStone2)
+
+        playingBoard.placeStone(playerStone4)
+        playingBoard.placeStone(playerStone5)
+
+        val actual = WinningRule().canPlace(playingBoard.board, playerStone3)
+        val expected = PlaceResult.Success.Finish(GameResult.WIN_BLACK)
+
+        assertThat(actual).isEqualTo(expected)
+    }
+}
